backend/presentation/http/api/auth/refresh.go
package auth

import (
	"encoding/json"
	"net/http"

	"github.com/khanzadimahdi/testproject/application/auth/refresh"
)

type refreshHandler struct {
	useCase *refresh.UseCase
}

func NewRefreshHandler(useCase *refresh.UseCase) *refreshHandler {
	return &refreshHandler{
		useCase: useCase,
	}
}

func (h *refreshHandler) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
	var request refresh.Request
	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
		rw.WriteHeader(http.StatusBadRequest)
		return
	}

	response, err := h.useCase.Execute(&request)
	switch {
	case err != nil:
		rw.WriteHeader(http.StatusInternalServerError)
	case response != nil && len(response.ValidationErrors) > 0:
		rw.Header().Add("Content-Type", "application/json")
		rw.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(rw).Encode(response)
	default:
		rw.Header().Add("Content-Type", "application/json")
		rw.WriteHeader(http.StatusOK)
		json.NewEncoder(rw).Encode(response)
	}
}
